Drugačen izpisa napak

+++ Dodajanje.php
@@ -83,7 +83,7 @@ if(isset($_POST['tabela'])){
                         $id = substr($podatekpostSQL, ($idNahaja+3));
                         array_push($podatkiZaPoslat, array($tabele[$i] => $id));
                     }
-                    if($tabela == "Izdelek" && $tabele[$i] == "Cena"){
+                    else if($tabela == "Izdelek" && $tabele[$i] == "Cena"){
                         $cena = str_replace(",", ".", $podatekpostSQL);
                         array_push($podatkiZaPoslat, array($tabele[$i] => $cena));
 
@@ -346,13 +346,13 @@ function Izjeme($tabela, $stolpec){
                                                 else if($vrstica['Field'] == "Cas_Enkrat" && $tabela == "Nacrtovani_Prevzemi"){
                                                     if(isset($_SESSION['temp'][$vrstica['Field']])){
                                                         echo "<div class='formvnosItem' style='display:flex; flex-direction: column; align-items: center;'>
-                                                        <div class='vnosNaslov'>". str_replace("_", " ", $vrstica['Field']).":</div>
+                                                        <div class='vnosNaslov'>Čas Enkrat:</div>
                                                         <input type='date' name='". $vrstica['Field'] ."' class='ipPB' value='". $_SESSION['temp'][$vrstica['Field']] ."'>
                                                         </div>";
                                                     }
                                                     else{
                                                         echo "<div class='formvnosItem' style='display:flex; flex-direction: column; align-items: center;'>
-                                                        <div class='vnosNaslov'>". str_replace("_", " ", $vrstica['Field']).":</div>
+                                                        <div class='vnosNaslov'>Datum za enkratni prevzem:</div>
                                                         <input type='date' name='". $vrstica['Field'] ."' class='ipPB'>
                                                         </div>";
                                                     }
@@ -409,12 +409,26 @@ function Izjeme($tabela, $stolpec){
                                                     $rezultatIzdelek = mysqli_query($povezava, $sql);
 
                                                     if(mysqli_num_rows($rezultatIzdelek) > 0){
+
+                                                        $sessionIzdelek = false;
+
+                                                        //Če je shranjen izdelk v sessionu da vrednost v sessionIzdelek
+                                                        if(isset($_SESSION['temp'][$vrstica['Field']])){
+                                                            $sessionIzdelek = $_SESSION['temp'][$vrstica['Field']];
+                                                        }
+
                                                         echo "<div class='formvnosItem' style='display:flex; flex-direction: column; align-items: center;'>
                                                                 <div class='vnosNaslov'>". str_replace("_", " ", $vrstica['Field']).":</div>
                                                                 <select name='". $vrstica['Field'] ."' class='select'>";
                                                     
                                                         while($vrsticaIzdelek = mysqli_fetch_assoc($rezultatIzdelek)){
-                                                            echo "<option value='" . $vrsticaIzdelek["Izdelek"] . "'>". $vrsticaIzdelek["Izdelek"] ."</option>";
+                                                            //Če je seessionIzdelek enak trenutnemu izdelku potem bo ta izbran
+                                                            if($sessionIzdelek == $vrsticaIzdelek["Izdelek"]){
+                                                                echo "<option value='" . $vrsticaIzdelek["Izdelek"] . "' selected>". $vrsticaIzdelek["Izdelek"] ."</option>";
+                                                            }
+                                                            else{
+                                                                echo "<option value='" . $vrsticaIzdelek["Izdelek"] . "'>". $vrsticaIzdelek["Izdelek"] ."</option>";
+                                                            }
                                 
                                                             
                                                         }
@@ -429,13 +443,28 @@ function Izjeme($tabela, $stolpec){
                                                     $rezultatStranka = mysqli_query($povezava, $sql);
 
                                                     if(mysqli_num_rows($rezultatStranka) > 0){
+
+                                                        $sessionStranka = false;
+
+                                                        //Če je shranjena stranka v sessionu da vrednost v sessionStranka
+                                                        if(isset($_SESSION['temp'][$vrstica['Field']])){
+                                                            $sessionStranka = $_SESSION['temp'][$vrstica['Field']];
+                                                        }
+
                                                         echo "<div class='formvnosItem'>";
                                                         echo "<div class='vnosNaslov'>Stranka:</div>";
                                                         echo "<input list='Stranke' name='". $vrstica['Field'] ."' id='Stranka'/>";
                                                         echo "<datalist id='Stranke'>";
 
                                                         while($vrsticaStranka = mysqli_fetch_assoc($rezultatStranka)){
-                                                            echo "<option value='" . $vrsticaStranka['Priimek'] . " " . $vrsticaStranka['Ime'] . " - " . $vrsticaStranka['id_stranke'] . "'>";
+
+                                                            //Če je seessionStranka enak trenutnemu izdelku potem bo ta izbran
+                                                            if($sessionStranka == $vrsticaStranka['id_stranke']){
+                                                                echo "<option value='" . $vrsticaStranka['Priimek'] . " " . $vrsticaStranka['Ime'] . " - " . $vrsticaStranka['id_stranke'] . "' selected>";
+                                                            }
+                                                            else{
+                                                                echo "<option value='" . $vrsticaStranka['Priimek'] . " " . $vrsticaStranka['Ime'] . " - " . $vrsticaStranka['id_stranke'] . "'>";
+                                                            }
                                                         }
                                                         echo "</datalist>";
                                                         echo "</div>";
@@ -450,13 +479,27 @@ function Izjeme($tabela, $stolpec){
                                                     $rezultatPosta = mysqli_query($povezava, $sql);
 
                                                     if(mysqli_num_rows($rezultatPosta) > 0){
+
+                                                        $sessionPosta = false;
+
+                                                        //Če je shranjena posta v sessionu da vrednost v sessionPosta
+                                                        if(isset($_SESSION['temp'][$vrstica['Field']])){
+                                                            $sessionPosta = $_SESSION['temp'][$vrstica['Field']];
+                                                        }
+
                                                         echo "<div class='formvnosItem'>";
                                                         echo "<div class='vnosNaslov'>Pošta:</div>";
                                                         echo "<input list='Postalist' name='". $vrstica['Field'] ."' id='Posta'/>";
                                                         echo "<datalist id='Postalist'>";
 
                                                         while($vrsticaPosta = mysqli_fetch_assoc($rezultatPosta)){
-                                                            echo "<option value='" . $vrsticaPosta['Postana_stevilka'] . " - " . $vrsticaPosta['Kraj'] . "'>";
+                                                            //Če je sessionPosta enak trenutni posti potem bo ta izbran
+                                                            if($sessionPosta == $vrsticaPosta['Postana_stevilka']){
+                                                                echo "<option value='" . $vrsticaPosta['Postana_stevilka'] . " - " . $vrsticaPosta['Kraj'] . "' selected>";
+                                                            }
+                                                            else{
+                                                                echo "<option value='" . $vrsticaPosta['Postana_stevilka'] . " - " . $vrsticaPosta['Kraj'] . "'>";
+                                                            }
                                                         }
                                                         echo "</datalist>";
                                                         echo "</div>";
@@ -470,11 +513,30 @@ function Izjeme($tabela, $stolpec){
                                                         for($i = 0; $i < count($omejitve); $i++){
                                                             if(isset($omejitve[$i][$vrstica['Field']])){
 
-                                                                echo "<div class='formvnosItem' style='display:flex; flex-direction: column; align-items: center;'>
-                                                                <div class='vnosNaslov'>". str_replace("_", " ", $vrstica['Field']).":</div>
-                                                                <select name='". $vrstica['Field'] ."' class='select'>";
+                                                                $sessionVrednost = false;
+
+                                                                //Če je shranjena vrednost v sessionu da vrednost v sessionVrednost
+                                                                if(isset($_SESSION['temp'][$vrstica['Field']])){
+                                                                    $sessionVrednost = $_SESSION['temp'][$vrstica['Field']];
+                                                                }
+
+                                                                echo "<div class='formvnosItem' style='display:flex; flex-direction: column; align-items: center;'>";
+                                                                if($vrstica['Field'] == "Cas"){
+                                                                    echo "<div class='vnosNaslov'>Čas:</div>";
+                                                                }
+                                                                else{
+                                                                    echo "<div class='vnosNaslov'>". str_replace("_", " ", $vrstica['Field']).":</div>";
+                                                                }
+                                                                
+                                                                echo "<select name='". $vrstica['Field'] ."' class='select'>";
                                                                     foreach($omejitve[$i][$vrstica['Field']] as $omejitev){
-                                                                        echo "<option value='$omejitev'>$omejitev</option>";
+                                                                        //Če je sessionPosta enak trenutni posti potem bo ta izbran
+                                                                        if($sessionVrednost == $omejitev){
+                                                                            echo "<option value='$omejitev' selected>$omejitev</option>";
+                                                                        }
+                                                                        else{
+                                                                            echo "<option value='$omejitev'>$omejitev</option>";
+                                                                        }
                                                                     }
                                                                     
 
@@ -536,20 +598,46 @@ function Izjeme($tabela, $stolpec){
                                 </div> 
 
                                 <?php 
+                                    //Dobi napako
                                     if(isset($_GET['napaka'])){
 
-                                        if(isset($tabele[$_GET['napaka']])){
-                                            if($tabela == "Prodaja" && $tabele[$_GET['napaka']] == "id_stranke"){
-                                                echo "<div class='napaka'>Vpišite veljavno Stranko</div>";
+                                        //Pogleda če obstajaja sporocila.php, če so jih includa in uporabi, drugače izpiše default
+                                        if(file_exists("Sporocila.php")){     
+
+                                            //Preveri, če obstaja stolpec za to napako, če uporabi sporocila.php drugače jo samo izpiše
+                                            if(isset($tabele[$_GET['napaka']])){
+
+                                                define('LahkoSporocila', TRUE);
+                                                include("Sporocila.php");
+
+                                                //Vrne true če ni prepoznana naoaka, drugače samo izpiše napako
+                                                if(NapakaSporocilo($tabela, $tabele[$_GET['napaka']])){
+                                                    $napaka = str_replace ( '%20', ' ', $_GET['napaka']);
+                                                    echo "<div class='napaka'>$napaka</div>";
+                                                }
+
                                             }
                                             else{
-                                                echo "<div class='napaka'>Vpišite veljavno ". str_replace("_", " ", $tabele[$_GET['napaka']]) ."</div>";
+                                                $napaka = str_replace ( '%20', ' ', $_GET['napaka']);
+                                                echo "<div class='napaka'>$napaka</div>";
                                             }
                                         }
                                         else{
-                                            $napaka = str_replace ( '%20', ' ', $_GET['napaka']);
-                                            echo "<div class='napaka'>$napaka</div>";
+                                            if(isset($tabele[$_GET['napaka']])){
+                                                if($tabela == "Prodaja" && $tabele[$_GET['napaka']] == "id_stranke"){
+                                                    echo "<div class='napaka'>Vpišite veljavno Stranko</div>";
+                                                }
+                                                else{
+                                                    echo "<div class='napaka'>Vpišite veljavno ". str_replace("_", " ", $tabele[$_GET['napaka']]) ."</div>";
+                                                }
+                                            }
+                                            else{
+                                                $napaka = str_replace ( '%20', ' ', $_GET['napaka']);
+                                                echo "<div class='napaka'>$napaka</div>";
+                                            }
                                         }
+
+                                        
 
                                     }
